refactor(llm): sseEvents carries the SSE parse scaffold for both chat adapters
The OpenAI-compatible and OpenRouter chat streams each stood up their own
`Sse.makeParser`, their own drain buffer, and the same decode → feed →
flatten → take-until-`[DONE]` pipeline. Only the malformed-output message
differed, and the two comments explaining the uncapped `maxEventSize` and the
ignored reconnect hint had drifted into different wordings of the same fact.

Move the scaffold to `sseEvents(bytes, malformedMessage)` and state those two
policies once, where the next adapter will read them.
