+++ packages/llm/src/openaiChat.ts
@@ -1,6 +1,5 @@
 // Third-party imports
 import { Cause, Effect, Exit, Stream } from 'effect';
-import { Sse } from 'effect/unstable/encoding';
 import OpenAI from 'openai';
 import { z } from 'zod';
 
@@ -13,6 +12,7 @@ import {
   ModelError,
   enrichModelError,
   pullStream,
+  sseEvents,
   readerAbortSignal,
   ResolvedTurnSchema,
   TurnRequestSchema,
@@ -1078,35 +1078,9 @@ export function openaiChatModel(
           >();
 
           const bytes = pullStream(() => body.read(), openaiFailure);
-          let parsedEvents: Sse.Event[] = [];
-          const parser = Sse.makeParser(
-            (event) => {
-              // Retry is only a reconnect hint; this operation never reconnects.
-              if (event._tag === 'Event') parsedEvents.push(event);
-            },
-            {
-              // Preserve the prior no-added-cap policy, not a bounded-memory claim.
-              maxEventSize: Number.POSITIVE_INFINITY,
-            },
-          );
-          const chunks = bytes.pipe(
-            Stream.decodeText,
-            Stream.mapEffect((text) =>
-              Effect.gen(function* () {
-                parsedEvents = [];
-                const failure = parser.feed(text);
-                if (failure !== undefined) {
-                  return yield* new ModelError({
-                    kind: 'malformed-output',
-                    message: 'The model returned malformed server-sent events.',
-                    cause: failure,
-                  });
-                }
-                return parsedEvents;
-              }),
-            ),
-            Stream.flattenIterable,
-            Stream.takeUntil((event) => event.data === '[DONE]'),
+          const chunks = sseEvents(
+            bytes,
+            'The model returned malformed server-sent events.',
           );
 
           const progress = chunks.pipe(
